requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via langchain-community
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   openai
attrs==25.3.0
    # via aiohttp
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
colorama==0.4.6
    # via tqdm
dataclasses-json==0.6.7
    # via langchain-community
distro==1.9.0
    # via openai
faiss-cpu==1.12.0
    # via -r requirements.in
filelock==3.19.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.7.0
    # via
    #   huggingface-hub
    #   torch
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langsmith
    #   openai
httpx-sse==0.4.1
    # via langchain-community
huggingface-hub==0.34.4
    # via
    #   langchain-huggingface
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.6
    # via torch
jiter==0.10.0
    # via openai
joblib==1.5.2
    # via scikit-learn
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.27
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.3.29
    # via
    #   -r requirements.in
    #   langchain-experimental
langchain-core==0.3.75
    # via
    #   langchain
    #   langchain-community
    #   langchain-experimental
    #   langchain-huggingface
    #   langchain-openai
    #   langchain-text-splitters
langchain-experimental==0.3.4
    # via -r requirements.in
langchain-huggingface==0.3.1
    # via -r requirements.in
langchain-openai==0.3.32
    # via -r requirements.in
langchain-text-splitters==0.3.9
    # via langchain
langsmith==0.4.20
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
markupsafe==3.0.2
    # via jinja2
marshmallow==3.26.1
    # via dataclasses-json
mpmath==1.3.0
    # via sympy
multidict==6.6.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
networkx==3.5
    # via torch
numpy==2.3.2
    # via
    #   faiss-cpu
    #   langchain-community
    #   pandas
    #   rank-bm25
    #   scikit-learn
    #   scipy
    #   transformers
openai==1.102.0
    # via langchain-openai
orjson==3.11.3
    # via langsmith
packaging==25.0
    # via
    #   faiss-cpu
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   marshmallow
    #   transformers
pandas==2.3.2
    # via -r requirements.in
pillow==11.3.0
    # via sentence-transformers
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
pydantic==2.11.7
    # via
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via langchain-community
pymupdf==1.26.4
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   pydantic-settings
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
    #   transformers
rank-bm25==0.2.2
    # via -r requirements.in
regex==2025.7.34
    # via
    #   tiktoken
    #   transformers
requests==2.32.5
    # via
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
    #   transformers
requests-toolbelt==1.0.0
    # via langsmith
safetensors==0.6.2
    # via transformers
scikit-learn==1.7.1
    # via sentence-transformers
scipy==1.16.1
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==5.1.0
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.43
    # via
    #   langchain
    #   langchain-community
sympy==1.14.0
    # via torch
tenacity==9.1.2
    # via
    #   langchain-community
    #   langchain-core
threadpoolctl==3.6.0
    # via scikit-learn
tiktoken==0.11.0
    # via langchain-openai
tokenizers==0.21.4
    # via
    #   langchain-huggingface
    #   transformers
torch==2.8.0
    # via sentence-transformers
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
    #   sentence-transformers
    #   transformers
transformers==4.55.4
    # via sentence-transformers
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   huggingface-hub
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sentence-transformers
    #   sqlalchemy
    #   torch
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
yarl==1.20.1
    # via aiohttp
zstandard==0.24.0
    # via langsmith
